+++ app/client_secret_575823433713-v5ehsvktk1ibsm1osnp8lf8mf4u8pq74.apps.googleusercontent.com.json
@@ -1,0 +1,1 @@
+{"installed":{"client_id":"575823433713-v5ehsvktk1ibsm1osnp8lf8mf4u8pq74.apps.googleusercontent.com","project_id":"notesapp-52e89","auth_uri":"https://accounts.google.com/o/oauth2/auth","token_uri":"https://oauth2.googleapis.com/token","auth_provider_x509_cert_url":"https://www.googleapis.com/oauth2/v1/certs","redirect_uris":["urn:ietf:wg:oauth:2.0:oob","http://localhost"]}}
