--- app/credentials.json
@@ -1,1 +1,0 @@
-{"web":{"client_id":"640130630116-8s0r3juoed6ba571281hug62q5o5dh7c.apps.googleusercontent.com","project_id":"goolog-1645465312326","auth_uri":"https://accounts.google.com/o/oauth2/auth","token_uri":"https://oauth2.googleapis.com/token","auth_provider_x509_cert_url":"https://www.googleapis.com/oauth2/v1/certs","client_secret":"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"}}
