--- app/src/main/java/com/homework/homework_6/observer/EventListener.java
@@ -1,8 +1,0 @@
-package com.homework.homework_6.observer;
-
-import com.homework.homework_6.data.CardData;
-
-public interface EventListener {
-    void updateData (CardData cardData);
-}
-
