+++ app/src/main/java/com/my/notes/data/Login.java
@@ -1,4 +1,4 @@
-package com.homework.homework_6.data;
+package com.my.notes.data;
 
 public interface Login {
     String login = "KEY";
