+++ app/src/main/java/com/my/notes/observer/EventListener.java
@@ -1,0 +1,8 @@
+package com.my.notes.observer;
+
+import com.my.notes.data.CardData;
+
+public interface EventListener {
+    void updateData (CardData cardData);
+}
+
