+++ app/src/main/java/com/my/notes/ui/ExitDialogFragment.java
@@ -1,4 +1,4 @@
-package com.homework.homework_6.ui;
+package com.my.notes.ui;
 
 import android.app.Dialog;
 import android.os.Bundle;
@@ -9,10 +9,9 @@ import androidx.annotation.Nullable;
 import androidx.appcompat.app.AlertDialog;
 import androidx.fragment.app.DialogFragment;
 
-import com.homework.homework_6.MainActivity;
+import com.my.notes.MainActivity;
 
-public class NoteDialogFragment extends DialogFragment {
-
+public class ExitDialogFragment extends DialogFragment {
     @NonNull
     @Override
     public Dialog onCreateDialog(@Nullable Bundle savedInstanceState) {
@@ -20,10 +19,14 @@ public class NoteDialogFragment extends DialogFragment {
         AlertDialog.Builder builder = new AlertDialog.Builder(activity);
         builder.setTitle("Внимание!")
                 .setCancelable(true)
-                .setMessage("Удалить заметку?")
-                .setPositiveButton("Да",(dialog,id)-> Toast.makeText(activity, "Заметка удалена!", Toast.LENGTH_SHORT).show())
+                .setMessage("Вы действительно хотитие выйти?")
+                .setPositiveButton("Да",(dialog,id)->{
+                    Toast.makeText(activity, "До свидания!", Toast.LENGTH_SHORT).show();
+    requireActivity().finish();
+    })
                 .setNegativeButton("Нет",(dialog,id)->{
                 });
-        return builder.create();
+        Dialog dialog = builder.create();
+        return dialog;
     }
 }
