+++ app/src/main/java/com/my/notes/ui/NoteDialogFragment.java
@@ -1,4 +1,4 @@
-package com.homework.homework_6.ui;
+package com.my.notes.ui;
 
 import android.app.Dialog;
 import android.os.Bundle;
@@ -9,9 +9,10 @@ import androidx.annotation.Nullable;
 import androidx.appcompat.app.AlertDialog;
 import androidx.fragment.app.DialogFragment;
 
-import com.homework.homework_6.MainActivity;
+import com.my.notes.MainActivity;
 
-public class ExitDialogFragment extends DialogFragment {
+public class NoteDialogFragment extends DialogFragment {
+
     @NonNull
     @Override
     public Dialog onCreateDialog(@Nullable Bundle savedInstanceState) {
@@ -19,14 +20,10 @@ public class ExitDialogFragment extends DialogFragment {
         AlertDialog.Builder builder = new AlertDialog.Builder(activity);
         builder.setTitle("Внимание!")
                 .setCancelable(true)
-                .setMessage("Вы действительно хотитие выйти?")
-                .setPositiveButton("Да",(dialog,id)->{
-                    Toast.makeText(activity, "До свидания!", Toast.LENGTH_SHORT).show();
-    requireActivity().finish();
-    })
+                .setMessage("Удалить заметку?")
+                .setPositiveButton("Да",(dialog,id)-> Toast.makeText(activity, "Заметка удалена!", Toast.LENGTH_SHORT).show())
                 .setNegativeButton("Нет",(dialog,id)->{
                 });
-        Dialog dialog = builder.create();
-        return dialog;
+        return builder.create();
     }
 }
