+++ app/src/main/java/com/my/notes/ui/NoteFragment.java
@@ -1,4 +1,4 @@
-package com.homework.homework_6.ui;
+package com.my.notes.ui;
 
 import static android.app.Activity.RESULT_OK;
 
@@ -21,11 +21,11 @@ import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.google.android.material.textfield.TextInputEditText;
 import com.google.android.material.textview.MaterialTextView;
 import com.google.firebase.firestore.FirebaseFirestore;
-import com.homework.homework_6.MainActivity;
-import com.homework.homework_6.R;
-import com.homework.homework_6.data.CardData;
-import com.homework.homework_6.data.Login;
-import com.homework.homework_6.observer.EventManager;
+import com.my.notes.MainActivity;
+import com.my.notes.R;
+import com.my.notes.data.CardData;
+import com.my.notes.data.Login;
+import com.my.notes.observer.EventManager;
 import java.text.SimpleDateFormat;
 import java.util.Calendar;
 import java.util.Date;
