+++ app/src/main/java/com/my/notes/ui/SettingsFragment.java
@@ -1,4 +1,4 @@
-package com.homework.homework_6.ui;
+package com.my.notes.ui;
 
 import android.os.Bundle;
 
@@ -12,7 +12,7 @@ import android.view.ViewGroup;
 import android.widget.Toast;
 
 import com.google.android.material.switchmaterial.SwitchMaterial;
-import com.homework.homework_6.R;
+import com.my.notes.R;
 
 
 public class SettingsFragment extends Fragment {
